move routes to App according to code-review
