+++ src/App.tsx
@@ -1,41 +1,35 @@
 import { useEffect } from 'react';
-import { Header } from './components/Header/Header';
-import styles from './App.module.scss';
-import { Outlet, useNavigate } from 'react-router-dom';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { fetchCourses } from './store/asyncActions';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import { AppDispatch } from './store';
-import { getIsAuth } from './store/user/selectors';
+import { Login } from './components/Login/Login';
+import { Courses } from './components/Courses/Courses';
+import { CreateCourse } from './components/CreateCourse/CreateCourse';
+import { CourseInfo } from './components/CourseInfo/CourseInfo';
+import { Registration } from './components/Registration/Registration';
+import Layout from './components/Layout/Layout';
 
 const App = () => {
-	const navigate = useNavigate();
-
 	const dispatch = useDispatch<AppDispatch>();
-	const isAuth = useSelector(getIsAuth);
 
 	useEffect(() => {
 		dispatch(fetchCourses());
 	}, [dispatch]);
 
-	useEffect(() => {
-		const isToken = localStorage.getItem('AUTH_TOKEN_REACT_COURSE');
-		if (isAuth && isToken && isToken.includes('Bearer')) {
-			navigate('/courses', { replace: true });
-		}
-		if (!isAuth) {
-			navigate('/login', { replace: true });
-		}
-	}, [isAuth]);
-
 	return (
-		<div>
-			<Header />
-			<div className={styles.container}>
-				<div>
-					<Outlet />
-				</div>
-			</div>
-		</div>
+		<BrowserRouter>
+			<Layout>
+				<Routes>
+					<Route path='login' element={<Login />} />
+					<Route path='courses' element={<Courses />} />
+					<Route path='courses/add' element={<CreateCourse />} />
+					<Route path='courses/:courseId/edit' element={<CreateCourse />} />
+					<Route path='courses/:courseId' element={<CourseInfo />} />
+					<Route path='registration' element={<Registration />} />
+				</Routes>
+			</Layout>
+		</BrowserRouter>
 	);
 };
